+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
     buildToolsVersion "25.0.3"
 
     defaultConfig {
-        applicationId "yamblz.voltek.com.weather"
+        applicationId "com.yamblz.voltek.weather"
         minSdkVersion 21
         targetSdkVersion 25
         versionCode 1
