+++ app/src/main/java/com/yamblz/voltek/weather/presentation/base/BaseFragment.java
@@ -1,4 +1,4 @@
-package yamblz.voltek.com.weather.presentation.base;
+package com.yamblz.voltek.weather.presentation.base;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
