+++ app/src/main/java/com/yamblz/voltek/weather/presentation/ui/about/AboutFragment.java
@@ -1,4 +1,4 @@
-package yamblz.voltek.com.weather.presentation.ui.settings;
+package com.yamblz.voltek.weather.presentation.ui.about;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -6,16 +6,16 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import yamblz.voltek.com.weather.R;
-import yamblz.voltek.com.weather.presentation.base.BaseFragment;
+import com.yamblz.voltek.weather.R;
+import com.yamblz.voltek.weather.presentation.base.BaseFragment;
 
-public class SettingsFragment extends BaseFragment {
+public class AboutFragment extends BaseFragment {
 
-    public static SettingsFragment newInstance() {
+    public static AboutFragment newInstance() {
 
         Bundle args = new Bundle();
 
-        SettingsFragment fragment = new SettingsFragment();
+        AboutFragment fragment = new AboutFragment();
         fragment.setArguments(args);
         return fragment;
     }
@@ -23,13 +23,13 @@ public class SettingsFragment extends BaseFragment {
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_settings, container, false);
+        return inflater.inflate(R.layout.fragment_about, container, false);
     }
 
     @Override
     public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        getActivity().setTitle(R.string.title_settings);
+        getActivity().setTitle(R.string.title_about);
     }
 }
