+++ app/src/main/java/com/yamblz/voltek/weather/presentation/ui/forecast/ForecastFragment.java
@@ -1,4 +1,4 @@
-package yamblz.voltek.com.weather.presentation.ui.about;
+package com.yamblz.voltek.weather.presentation.ui.forecast;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -6,16 +6,16 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import yamblz.voltek.com.weather.R;
-import yamblz.voltek.com.weather.presentation.base.BaseFragment;
+import com.yamblz.voltek.weather.R;
+import com.yamblz.voltek.weather.presentation.base.BaseFragment;
 
-public class AboutFragment extends BaseFragment {
+public class ForecastFragment extends BaseFragment {
 
-    public static AboutFragment newInstance() {
+    public static ForecastFragment newInstance() {
 
         Bundle args = new Bundle();
 
-        AboutFragment fragment = new AboutFragment();
+        ForecastFragment fragment = new ForecastFragment();
         fragment.setArguments(args);
         return fragment;
     }
@@ -23,13 +23,13 @@ public class AboutFragment extends BaseFragment {
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_about, container, false);
+        return inflater.inflate(R.layout.fragment_forecast, container, false);
     }
 
     @Override
     public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        getActivity().setTitle(R.string.title_about);
+        getActivity().setTitle(R.string.title_forecast);
     }
 }
