+++ app/src/main/java/com/yamblz/voltek/weather/presentation/ui/settings/SettingsFragment.java
@@ -1,4 +1,4 @@
-package yamblz.voltek.com.weather.presentation.ui.forecast;
+package com.yamblz.voltek.weather.presentation.ui.settings;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -6,16 +6,16 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import yamblz.voltek.com.weather.R;
-import yamblz.voltek.com.weather.presentation.base.BaseFragment;
+import com.yamblz.voltek.weather.R;
+import com.yamblz.voltek.weather.presentation.base.BaseFragment;
 
-public class ForecastFragment extends BaseFragment {
+public class SettingsFragment extends BaseFragment {
 
-    public static ForecastFragment newInstance() {
+    public static SettingsFragment newInstance() {
 
         Bundle args = new Bundle();
 
-        ForecastFragment fragment = new ForecastFragment();
+        SettingsFragment fragment = new SettingsFragment();
         fragment.setArguments(args);
         return fragment;
     }
@@ -23,13 +23,13 @@ public class ForecastFragment extends BaseFragment {
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_forecast, container, false);
+        return inflater.inflate(R.layout.fragment_settings, container, false);
     }
 
     @Override
     public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        getActivity().setTitle(R.string.title_forecast);
+        getActivity().setTitle(R.string.title_settings);
     }
 }
